requirements.txt
opencv-contrib-python
numpy
pandas
torch
Pillow
torchvision
matplotlib